added parallelization for the range calculations
used std::for_each( std::execution::par, ... ) to run the different calculations in parallel.
added a lambda to wrap the existing code into a function and used the cap-group to work with the existing mappers.
A local mutex is used to collect all different calculations back inside the main execution path.